Remove `engine` constraints in package.json
I had a quick look into the documentation of these constraints: git@github.com:roschaefer/vue-izitoast.git
Regarding the `npm` constraints they say:
> this field is advisory onlyreasoning
so even they are not really confident on the usefulness of the constraints for `npm`. Can we just be optimistic and assume it works fine? If people get back to us with problems we can still add a constraint if we *know*  there are any issues.

I cannot see anything in the code of this package that would justify having these constraints.

This commit makes #20 obsolete.

If I missed anything here I would love if you could inform me about why we have the constraints at all.